+++ lib/rack/bad_request.rb
@@ -1,0 +1,4 @@
+module Rack
+  module BadRequest
+  end
+end